Rename pulpcore.app.logging module to prevent name shadowing
manage.py commands like "collectstatic" would fail due to circular
import as a result of confusing logging.py with the logging module in
stdlib.

[noissue]
