+++ pulpcore/app/loggers.py
@@ -1,4 +1,3 @@
 import logging
 
-
 deprecation_logger = logging.getLogger("pulpcore.deprecation")
